WIP Plugin API (currently broken); make album search a plugin
