+++ mps_youtube/__init__.py
@@ -1,2 +1,7 @@
 from .main import init
+from .main import __version__, __notes__, __author__, __license__, __url__
+from .config import Config
+from . import c
+from .plugin import Plugin, registerPlugin
+
 init()
